Home page API endpoints moved to api/home_page_api_views.py
